NodeStrap
NodeStrap is a unique implementation of nodejs, bootstrap and PHP.

It uses nodecrud, startbootsrap-agency template, customized for nodejs.
Built in support for PHP included in app.js.

Make sure you install: "npm install php" (if it is not being run
automatically) when you do "npm install" to this master directory.